+++ src/components/breadcrumbs/breadcrumbs.component.tsx
@@ -1,9 +1,10 @@
 import React from "react";
+import { SpaceProps } from "styled-system";
 import tagComponent, { TagProps } from "../../__internal__/utils/helpers/tags";
 import StyledBreadcrumbs from "./breadcrumbs.style";
 import useLocale from "../../hooks/__internal__/useLocale";
 
-export interface BreadcrumbsProps extends TagProps {
+export interface BreadcrumbsProps extends TagProps, SpaceProps {
   /** Child crumbs to display */
   children: React.ReactNode;
 }
@@ -17,6 +18,7 @@ export const Breadcrumbs = React.forwardRef<HTMLElement, BreadcrumbsProps>(
         role="navigation"
         {...tagComponent("breadcrumbs", rest)}
         aria-label={l.breadcrumbs.ariaLabel()}
+        {...rest}
       >
         <ol>{children}</ol>
       </StyledBreadcrumbs>