+++ src/components/button-toggle/button-toggle.stories.tsx
@@ -1,129 +1,104 @@
 import React from "react";
 import { ComponentStory } from "@storybook/react";
 
-import ButtonToggle from ".";
-import ButtonToggleGroup from "../button-toggle-group";
+import { ButtonToggle, ButtonToggleGroup } from ".";
 import Box from "../box";
 
 export const Default: ComponentStory<typeof ButtonToggle> = () => (
-  <ButtonToggleGroup id="button-toggle-group-id" name="options">
-    <ButtonToggle name="default" value="foo">
-      Foo
-    </ButtonToggle>
-    <ButtonToggle name="default" value="bar">
-      Bar
-    </ButtonToggle>
-    <ButtonToggle name="default" value="baz">
-      Baz
-    </ButtonToggle>
+  <ButtonToggleGroup id="button-toggle-group-id" label="Default example">
+    <ButtonToggle value="foo">Foo</ButtonToggle>
+    <ButtonToggle value="bar">Bar</ButtonToggle>
+    <ButtonToggle value="baz">Baz</ButtonToggle>
   </ButtonToggleGroup>
 );
 
 export const DefaultWrappedText: ComponentStory<typeof ButtonToggle> = () => (
   <Box margin={4} width="250px" display="flex" flexWrap="nowrap">
-    <ButtonToggle name="default" value="wraps" grouped>
-      Some text that wraps
-    </ButtonToggle>
-    <ButtonToggle name="default" value="foobar" grouped>
-      FooBar
-    </ButtonToggle>
+    <ButtonToggleGroup id="button-toggle-group-id" label="Wrapped text example">
+      <ButtonToggle value="wraps" grouped>
+        Some text that wraps
+      </ButtonToggle>
+      <ButtonToggle value="foobar" grouped>
+        FooBar
+      </ButtonToggle>
+    </ButtonToggleGroup>
   </Box>
 );
 
 export const DefaultSmallIcon: ComponentStory<typeof ButtonToggle> = () => (
-  <ButtonToggleGroup id="button-toggle-group-id" name="options">
-    <ButtonToggle name="sizes" value="foo" buttonIcon="add">
+  <ButtonToggleGroup id="button-toggle-group-id" label="Small icon example">
+    <ButtonToggle value="foo" buttonIcon="add">
       Add
     </ButtonToggle>
-    <ButtonToggle name="sizes" value="bar" buttonIcon="share">
+    <ButtonToggle value="bar" buttonIcon="share">
       Share
     </ButtonToggle>
-    <ButtonToggle name="sizes" value="baz" buttonIcon="tick">
+    <ButtonToggle value="baz" buttonIcon="tick">
       Tick
     </ButtonToggle>
   </ButtonToggleGroup>
 );
 
 export const DefaultLargeIcon: ComponentStory<typeof ButtonToggle> = () => (
-  <ButtonToggleGroup id="button-toggle-group-id" name="options">
-    <ButtonToggle
-      name="sizes"
-      value="foo"
-      buttonIcon="add"
-      buttonIconSize="large"
-    >
+  <ButtonToggleGroup id="button-toggle-group-id" label="Large icon example">
+    <ButtonToggle value="foo" buttonIcon="add" buttonIconSize="large">
       Add
     </ButtonToggle>
-    <ButtonToggle
-      name="sizes"
-      value="bar"
-      buttonIcon="share"
-      buttonIconSize="large"
-    >
+    <ButtonToggle value="bar" buttonIcon="share" buttonIconSize="large">
       Share
     </ButtonToggle>
-    <ButtonToggle
-      name="sizes"
-      value="baz"
-      buttonIcon="tick"
-      buttonIconSize="large"
-    >
+    <ButtonToggle value="baz" buttonIcon="tick" buttonIconSize="large">
       Tick
     </ButtonToggle>
   </ButtonToggleGroup>
 );
 
 export const iconOnly: ComponentStory<typeof ButtonToggle> = () => (
-  <ButtonToggleGroup id="button-toggle-group-id" name="options">
-    <ButtonToggle name="sizes" value="foo" buttonIcon="add" aria-label="add" />
-    <ButtonToggle
-      name="sizes"
-      value="bar"
-      buttonIcon="share"
-      aria-label="share"
-    />
-    <ButtonToggle
-      name="sizes"
-      value="baz"
-      buttonIcon="tick"
-      aria-label="tick"
-    />
+  <ButtonToggleGroup id="button-toggle-group-id" label="Icon only example">
+    <ButtonToggle value="foo" buttonIcon="add" aria-label="add" />
+    <ButtonToggle value="bar" buttonIcon="share" aria-label="share" />
+    <ButtonToggle value="baz" buttonIcon="tick" aria-label="tick" />
   </ButtonToggleGroup>
 );
 
 export const small: ComponentStory<typeof ButtonToggle> = () => (
-  <ButtonToggleGroup id="button-toggle-group-id" name="options">
-    <ButtonToggle size="small" name="sizes" value="foo">
+  <ButtonToggleGroup id="button-toggle-group-id" label="Small example">
+    <ButtonToggle size="small" value="foo">
       Add
     </ButtonToggle>
-    <ButtonToggle size="small" name="sizes" value="bar">
+    <ButtonToggle size="small" value="bar">
       Share
     </ButtonToggle>
-    <ButtonToggle size="small" name="sizes" value="baz">
+    <ButtonToggle size="small" value="baz">
       Tick
     </ButtonToggle>
   </ButtonToggleGroup>
 );
 
 export const smallSmallIcon: ComponentStory<typeof ButtonToggle> = () => (
-  <ButtonToggleGroup id="button-toggle-group-id" name="options">
-    <ButtonToggle size="small" name="sizes" value="foo" buttonIcon="add">
+  <ButtonToggleGroup
+    id="button-toggle-group-id"
+    label="Small with small icon example"
+  >
+    <ButtonToggle size="small" value="foo" buttonIcon="add">
       Add
     </ButtonToggle>
-    <ButtonToggle size="small" name="sizes" value="bar" buttonIcon="share">
+    <ButtonToggle size="small" value="bar" buttonIcon="share">
       Share
     </ButtonToggle>
-    <ButtonToggle size="small" name="sizes" value="baz" buttonIcon="tick">
+    <ButtonToggle size="small" value="baz" buttonIcon="tick">
       Tick
     </ButtonToggle>
   </ButtonToggleGroup>
 );
 
 export const smallLargeIcon: ComponentStory<typeof ButtonToggle> = () => (
-  <ButtonToggleGroup id="button-toggle-group-id" name="options">
+  <ButtonToggleGroup
+    id="button-toggle-group-id"
+    label="Small with large icon example"
+  >
     <ButtonToggle
       size="small"
-      name="sizes"
       value="foo"
       buttonIcon="add"
       buttonIconSize="large"
@@ -132,7 +107,6 @@ export const smallLargeIcon: ComponentStory<typeof ButtonToggle> = () => (
     </ButtonToggle>
     <ButtonToggle
       size="small"
-      name="sizes"
       value="bar"
       buttonIcon="share"
       buttonIconSize="large"
@@ -141,7 +115,6 @@ export const smallLargeIcon: ComponentStory<typeof ButtonToggle> = () => (
     </ButtonToggle>
     <ButtonToggle
       size="small"
-      name="sizes"
       value="baz"
       buttonIcon="tick"
       buttonIconSize="large"
@@ -152,38 +125,43 @@ export const smallLargeIcon: ComponentStory<typeof ButtonToggle> = () => (
 );
 
 export const large: ComponentStory<typeof ButtonToggle> = () => (
-  <ButtonToggleGroup id="button-toggle-group-id" name="options">
-    <ButtonToggle size="large" name="sizes" value="foo">
+  <ButtonToggleGroup id="button-toggle-group-id" label="Large example">
+    <ButtonToggle size="large" value="foo">
       Add
     </ButtonToggle>
-    <ButtonToggle size="large" name="sizes" value="bar">
+    <ButtonToggle size="large" value="bar">
       Share
     </ButtonToggle>
-    <ButtonToggle size="large" name="sizes" value="baz">
+    <ButtonToggle size="large" value="baz">
       Tick
     </ButtonToggle>
   </ButtonToggleGroup>
 );
 
 export const largeSmallIcon: ComponentStory<typeof ButtonToggle> = () => (
-  <ButtonToggleGroup id="button-toggle-group-id" name="options">
-    <ButtonToggle size="large" name="sizes" value="foo" buttonIcon="add">
+  <ButtonToggleGroup
+    id="button-toggle-group-id"
+    label="Large with small icon example"
+  >
+    <ButtonToggle size="large" value="foo" buttonIcon="add">
       Add
     </ButtonToggle>
-    <ButtonToggle size="large" name="sizes" value="bar" buttonIcon="share">
+    <ButtonToggle size="large" value="bar" buttonIcon="share">
       Share
     </ButtonToggle>
-    <ButtonToggle size="large" name="sizes" value="baz" buttonIcon="tick">
+    <ButtonToggle size="large" value="baz" buttonIcon="tick">
       Tick
     </ButtonToggle>
   </ButtonToggleGroup>
 );
 
 export const largeLargeIcon: ComponentStory<typeof ButtonToggle> = () => (
-  <ButtonToggleGroup id="button-toggle-group-id" name="options">
+  <ButtonToggleGroup
+    id="button-toggle-group-id"
+    label="Large with large icon example"
+  >
     <ButtonToggle
       size="large"
-      name="sizes"
       value="foo"
       buttonIcon="add"
       buttonIconSize="large"
@@ -192,7 +170,6 @@ export const largeLargeIcon: ComponentStory<typeof ButtonToggle> = () => (
     </ButtonToggle>
     <ButtonToggle
       size="large"
-      name="sizes"
       value="bar"
       buttonIcon="share"
       buttonIconSize="large"
@@ -201,7 +178,6 @@ export const largeLargeIcon: ComponentStory<typeof ButtonToggle> = () => (
     </ButtonToggle>
     <ButtonToggle
       size="large"
-      name="sizes"
       value="baz"
       buttonIcon="tick"
       buttonIconSize="large"
@@ -212,28 +188,28 @@ export const largeLargeIcon: ComponentStory<typeof ButtonToggle> = () => (
 );
 
 export const disabled: ComponentStory<typeof ButtonToggle> = () => (
-  <ButtonToggleGroup id="button-toggle-group-id" name="options">
-    <ButtonToggle name="disabled" value="foo" disabled>
+  <ButtonToggleGroup id="button-toggle-group-id" label="Disabled example">
+    <ButtonToggle value="foo" disabled>
       Foo
     </ButtonToggle>
-    <ButtonToggle name="disabled" value="bar" disabled>
+    <ButtonToggle value="bar" disabled>
       Bar
     </ButtonToggle>
-    <ButtonToggle name="disabled" value="baz" disabled>
+    <ButtonToggle value="baz" disabled>
       Baz
     </ButtonToggle>
   </ButtonToggleGroup>
 );
 
 export const grouped: ComponentStory<typeof ButtonToggle> = () => (
-  <ButtonToggleGroup id="button-toggle-group-id" name="grouped-options">
-    <ButtonToggle name="grouped" value="foo" grouped>
+  <ButtonToggleGroup id="button-toggle-group-id" label="Grouped example">
+    <ButtonToggle value="foo" grouped>
       Foo
     </ButtonToggle>
-    <ButtonToggle name="grouped" value="bar" grouped>
+    <ButtonToggle value="bar" grouped>
       Bar
     </ButtonToggle>
-    <ButtonToggle name="grouped" value="baz" grouped>
+    <ButtonToggle value="baz" grouped>
       Baz
     </ButtonToggle>
   </ButtonToggleGroup>